query-agent-backend/requirements.txt
google-generativeai
playwright
faiss-cpu
sentence-transformers
numpy
requests 